millionx-phase2/scrapers/requirements.txt
# MillionX Phase 2 - Scraper Dependencies
# Python 3.11+

# Kafka client
kafka-python==2.0.2

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# API clients
requests==2.31.0
ShopifyAPI==12.3.0

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Proxy rotation and anti-bot
proxy-requests==0.6.0
python-anticaptcha==1.0.0  # Optional CAPTCHA solving
fake-useragent==1.4.0

# Apify fallback (optional)
apify-client==1.3.0

# Utilities
python-dateutil==2.8.2

# Logging
colorlog==6.8.0
